+++ src/qac_impl/IncNgTrie/resultset.h
@@ -8,6 +8,8 @@
 #define __RESULTSET_H__
 
 //#include <unordered_set>
+#include "cstddef"
+#include <cstddef>
 #include <unordered_set>
 
 using namespace std;
@@ -16,10 +18,10 @@ namespace dbwsim{
 
 class ResultSet {
  public:
-  unordered_set<int> result_ids_;
+  unordered_set<size_t> result_ids_;  // Unni int -> size_t
   //set<int> result_ids_;
   
-  void add (int document_id) {
+  void add (size_t document_id) { // Unni int -> size_t
     result_ids_.insert(document_id);
   }
   
@@ -27,7 +29,7 @@ class ResultSet {
     result_ids_.clear();
   }
 
-  int size() {
+  size_t size() { // Unni int -> size_t
     return result_ids_.size();
   } 
 };
